docs(checkbox): add playground example
