+++ src/components/checkbox/checkbox-test.stories.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { Meta, StoryObj } from "@storybook/react-vite";
 import generateStyledSystemProps from "../../../.storybook/utils/styled-system-props";
 
-import { Checkbox, CheckboxProps } from ".";
+import { Checkbox, CheckboxGroup, CheckboxProps } from ".";
 import Box from "../box";
 import Textbox from "../textbox";
 import Icon from "../icon";
@@ -254,3 +254,142 @@ export const IndeterminateSizesWithFocus: Story = {
     },
   },
 };
+
+// Documentation regression stories moved from the public docs.
+
+const DocumentationControlledCheckbox = ({
+  ...args
+}: Omit<CheckboxProps, "checked" | "onChange">) => {
+  const [isChecked, setIsChecked] = useState(false);
+
+  return (
+    <Checkbox
+      checked={isChecked}
+      onChange={(e) => setIsChecked(e.target.checked)}
+      {...args}
+    />
+  );
+};
+
+export const WithLabel: Story = {
+  render: DocumentationControlledCheckbox,
+  args: {
+    label: "Checkbox",
+  },
+};
+
+export const WithInputHint: Story = {
+  ...WithLabel,
+  args: {
+    ...WithLabel.args,
+    inputHint: "Input Hint",
+  },
+};
+
+export const Sizes: Story = () => {
+  const [checkedSmall, setCheckedSmall] = useState(false);
+  const [checkedMedium, setCheckedMedium] = useState(false);
+  const [checkedLarge, setCheckedLarge] = useState(false);
+
+  return (
+    <Box display="flex" flexDirection="row" justifyContent="space-around">
+      <Checkbox
+        label="Small Checkbox"
+        size="small"
+        checked={checkedSmall}
+        onChange={() => {
+          setCheckedSmall(!checkedSmall);
+        }}
+      />
+      <Checkbox
+        label="Medium Checkbox"
+        size="medium"
+        checked={checkedMedium}
+        onChange={() => {
+          setCheckedMedium(!checkedMedium);
+        }}
+      />
+      <Checkbox
+        label="Large Checkbox"
+        size="large"
+        checked={checkedLarge}
+        onChange={() => {
+          setCheckedLarge(!checkedLarge);
+        }}
+      />
+    </Box>
+  );
+};
+Sizes.storyName = "Sizes";
+
+export const IndeterminateState: Story = () => {
+  const [items, setItems] = useState([
+    { id: "checkbox-1", label: "Checkbox 1", checked: true },
+    { id: "checkbox-2", label: "Checkbox 2", checked: false },
+    { id: "checkbox-3", label: "Checkbox 3", checked: false },
+  ]);
+
+  const checkedCount = items.filter((item) => item.checked).length;
+  const allChecked = checkedCount === items.length;
+  const someChecked = checkedCount > 0 && checkedCount < items.length;
+  const controlledIds = items.map((item) => item.id).join(" ");
+
+  const handleSelectAll = () => {
+    const newChecked = !allChecked;
+    setItems(items.map((item) => ({ ...item, checked: newChecked })));
+  };
+
+  const handleChange = (id: string, checked: boolean) => {
+    setItems(
+      items.map((item) => (item.id === id ? { ...item, checked } : item)),
+    );
+  };
+
+  return (
+    <>
+      <Checkbox
+        label="Select All"
+        indeterminate={someChecked}
+        checked={allChecked}
+        onChange={handleSelectAll}
+        aria-controls={controlledIds}
+      />
+      <CheckboxGroup m={2}>
+        {items.map((item) => (
+          <Checkbox
+            key={item.id}
+            id={item.id}
+            label={item.label}
+            checked={item.checked}
+            onChange={(ev) => handleChange(item.id, ev.target.checked)}
+          />
+        ))}
+      </CheckboxGroup>
+    </>
+  );
+};
+IndeterminateState.storyName = "Indeterminate State";
+
+export const Required: Story = {
+  ...WithLabel,
+  args: {
+    ...WithLabel.args,
+    required: true,
+  },
+};
+
+export const Disabled: Story = {
+  ...WithInputHint,
+  args: {
+    ...WithInputHint.args,
+    required: true,
+    disabled: true,
+  },
+};
+
+WithLabel.parameters = { chromatic: { disableSnapshot: true } };
+WithInputHint.parameters = { chromatic: { disableSnapshot: true } };
+Sizes.parameters = { chromatic: { disableSnapshot: true } };
+IndeterminateState.parameters = { chromatic: { disableSnapshot: true } };
+Required.parameters = { chromatic: { disableSnapshot: true } };
+Disabled.parameters = { chromatic: { disableSnapshot: true } };
