+++ src/components/checkbox/checkbox.stories.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { Meta, StoryObj } from "@storybook/react-vite";
-import { Checkbox, CheckboxProps, CheckboxGroup } from ".";
+import { Checkbox, CheckboxProps } from ".";
 import Box from "../box";
 import Textbox from "../textbox";
 import Icon from "../icon";
@@ -15,6 +15,19 @@ const meta = {
   component: Checkbox,
   argTypes: {
     ...styledSystemProps,
+    label: {
+      control: "text",
+    },
+    disabled: {
+      control: "boolean",
+    },
+    required: {
+      control: "boolean",
+    },
+    size: {
+      options: ["small", "medium", "large"],
+      control: { type: "radio" },
+    },
   },
   parameters: {
     chromatic: { disableSnapshot: true },
@@ -26,6 +39,30 @@ const meta = {
 
 export default meta;
 type Story = StoryObj<typeof Checkbox>;
+
+export const Playground: Story = {
+  render: (args) => {
+    const [isChecked, setIsChecked] = useState(false);
+    return (
+      <Checkbox
+        {...args}
+        checked={isChecked}
+        onChange={(e) => setIsChecked(e.target.checked)}
+      />
+    );
+  },
+  args: {
+    label: "Checkbox",
+    inputHint: "Hint text",
+    disabled: false,
+    required: false,
+    size: "medium",
+    indeterminate: false,
+    error: "",
+    progressiveDisclosure: "Additional content shown when checked",
+  },
+};
+Playground.storyName = "Playground";
 
 const ControlledCheckbox = ({
   ...args
@@ -41,57 +78,6 @@ const ControlledCheckbox = ({
   );
 };
 
-export const WithLabel: Story = {
-  render: ControlledCheckbox,
-  args: {
-    label: "Checkbox",
-  },
-};
-
-export const WithInputHint: Story = {
-  ...WithLabel,
-  args: {
-    ...WithLabel.args,
-    inputHint: "Input Hint",
-  },
-};
-
-export const Sizes: Story = () => {
-  const [checkedSmall, setCheckedSmall] = useState(false);
-  const [checkedMedium, setCheckedMedium] = useState(false);
-  const [checkedLarge, setCheckedLarge] = useState(false);
-
-  return (
-    <Box display="flex" flexDirection="row" justifyContent="space-around">
-      <Checkbox
-        label="Small Checkbox"
-        size="small"
-        checked={checkedSmall}
-        onChange={() => {
-          setCheckedSmall(!checkedSmall);
-        }}
-      />
-      <Checkbox
-        label="Medium Checkbox"
-        size="medium"
-        checked={checkedMedium}
-        onChange={() => {
-          setCheckedMedium(!checkedMedium);
-        }}
-      />
-      <Checkbox
-        label="Large Checkbox"
-        size="large"
-        checked={checkedLarge}
-        onChange={() => {
-          setCheckedLarge(!checkedLarge);
-        }}
-      />
-    </Box>
-  );
-};
-Sizes.storyName = "Sizes";
-
 const DisclosedContent = () => {
   const [textboxValue, setTextboxValue] = useState("");
 
@@ -106,63 +92,6 @@ const DisclosedContent = () => {
   );
 };
 
-export const ProgressiveDisclosure: Story = {
-  ...WithLabel,
-  args: {
-    ...WithLabel.args,
-    checked: true,
-    progressiveDisclosure: <DisclosedContent />,
-  },
-};
-
-export const IndeterminateState: Story = () => {
-  const [items, setItems] = useState([
-    { id: "checkbox-1", label: "Checkbox 1", checked: true },
-    { id: "checkbox-2", label: "Checkbox 2", checked: false },
-    { id: "checkbox-3", label: "Checkbox 3", checked: false },
-  ]);
-
-  const checkedCount = items.filter((item) => item.checked).length;
-  const allChecked = checkedCount === items.length;
-  const someChecked = checkedCount > 0 && checkedCount < items.length;
-  const controlledIds = items.map((item) => item.id).join(" ");
-
-  const handleSelectAll = () => {
-    const newChecked = !allChecked;
-    setItems(items.map((item) => ({ ...item, checked: newChecked })));
-  };
-
-  const handleChange = (id: string, checked: boolean) => {
-    setItems(
-      items.map((item) => (item.id === id ? { ...item, checked } : item)),
-    );
-  };
-
-  return (
-    <>
-      <Checkbox
-        label="Select All"
-        indeterminate={someChecked}
-        checked={allChecked}
-        onChange={handleSelectAll}
-        aria-controls={controlledIds}
-      />
-      <CheckboxGroup m={2}>
-        {items.map((item) => (
-          <Checkbox
-            key={item.id}
-            id={item.id}
-            label={item.label}
-            checked={item.checked}
-            onChange={(ev) => handleChange(item.id, ev.target.checked)}
-          />
-        ))}
-      </CheckboxGroup>
-    </>
-  );
-};
-IndeterminateState.storyName = "Indeterminate State";
-
 const CustomLabel = () => (
   <>
     <Icon type="placeholder" aria-hidden />
@@ -170,26 +99,17 @@ const CustomLabel = () => (
   </>
 );
 
+export const ProgressiveDisclosure: Story = {
+  render: ControlledCheckbox,
+  args: {
+    label: "Checkbox",
+    progressiveDisclosure: <DisclosedContent />,
+  },
+};
+
 export const WithCustomLabel: Story = {
   render: ControlledCheckbox,
   args: {
     label: <CustomLabel />,
-  },
-};
-
-export const Required: Story = {
-  ...WithLabel,
-  args: {
-    ...WithLabel.args,
-    required: true,
-  },
-};
-
-export const Disabled: Story = {
-  ...WithInputHint,
-  args: {
-    ...WithInputHint.args,
-    required: true,
-    disabled: true,
   },
 };
